backend/requirements.txt
pyaudio==0.2.14
websocket-client==1.8.0
python-dotenv==1.0.1
requests==2.32.5
transformers==4.48.0
accelerate==1.1.1
torch==2.9.0
